requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
openai-whisper==20250625
torch==2.11.0
numpy==1.26.4